Remove cloudformation command flags in favour of subcommands

+++ cf/delete_stack.go
@@ -5,17 +5,23 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
+	"github.com/urfave/cli"
 )
 
-func deleteStack(name string) bool {
-	awsName := aws.String(name)
+func deleteStack(ctx *cli.Context) error {
+	name := ctx.Args().First()
+	if name == "" {
+		cli.ShowSubcommandHelp(ctx)
+		return nil
+	}
+
 	_, err := cfClient.DeleteStack(&cloudformation.DeleteStackInput{
-		StackName: awsName,
+		StackName: aws.String(name),
 	})
 	if err != nil {
 		log.Fatalln("ERROR", err)
 	}
 
-	printStackEvents(awsName)
-	return true
+	printStackEvents(name)
+	return nil
 }
